Each crse on own sheet, graphs have data issues

+++ project/src/test/java/edu/odu/cs/cs350/pne/TestMain.java
@@ -145,7 +145,7 @@ public class TestMain {
     @Test
     public void TestGetPercentagePassed() {
         int Total = 100, Passed = 27;
-        assertThat(new Main().GetPercentagePassed(Total, Passed), equalTo((float) 0.27));
+        assertThat(new Main().GetPercentagePassed(Total, Passed), equalTo((Double) 0.27));
     }
 
     @Test
